backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
flake8==5.0.4
flake8-mypy==17.8.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
promise==2.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
